GUI skeleton opgezet
