--- src/main/java/tudelft/ti2806/pl3/visualization/Application.java
@@ -1,88 +1,0 @@
-package tudelft.ti2806.pl3.visualization;
-
-import org.graphstream.graph.Graph;
-import org.graphstream.ui.swingViewer.View;
-import org.graphstream.ui.swingViewer.Viewer;
-import tudelft.ti2806.pl3.data.GraphParser;
-
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.*;
-import java.io.File;
-import java.io.FileNotFoundException;
-
-public class Application extends JDialog {
-    private JPanel contentPane;
-    private JButton buttonOK;
-    private JButton buttonCancel;
-    private JPanel graphView;
-
-    public Application() {
-        setContentPane(contentPane);
-        contentPane.setPreferredSize(new Dimension(1800, 1000));
-        setModal(true);
-        getRootPane().setDefaultButton(buttonOK);
-
-        buttonOK.addActionListener(new ActionListener() {
-            public void actionPerformed(ActionEvent e) {
-                onOK();
-            }
-        });
-
-        buttonCancel.addActionListener(new ActionListener() {
-            public void actionPerformed(ActionEvent e) {
-                onCancel();
-            }
-        });
-
-        // call onCancel() when cross is clicked
-        setDefaultCloseOperation(DO_NOTHING_ON_CLOSE);
-        addWindowListener(new WindowAdapter() {
-            public void windowClosing(WindowEvent e) {
-                onCancel();
-            }
-        });
-
-        // call onCancel() on ESCAPE
-        contentPane.registerKeyboardAction(new ActionListener() {
-            public void actionPerformed(ActionEvent e) {
-                onCancel();
-            }
-        }, KeyStroke.getKeyStroke(KeyEvent.VK_ESCAPE, 0), JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT);
-
-        Graph graph = null;
-        try {
-            graph = GraphParser.parseGraph("hello", new File(
-                    "data/10_strains_graph/simple_graph.node.graph"), new File(
-                    "data/10_strains_graph/simple_graph.edge.graph"));
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        }
-
-        graph.addAttribute("ui.antialias");
-
-        Viewer viewer = new Viewer(graph, Viewer.ThreadingModel.GRAPH_IN_ANOTHER_THREAD);
-        viewer.enableAutoLayout();
-
-        View view = viewer.addDefaultView(false);
-        graphView.setLayout(new BorderLayout());
-        graphView.add(view);
-    }
-
-    private void onOK() {
-        // add your code here
-        dispose();
-    }
-
-    private void onCancel() {
-        // add your code here if necessary
-        dispose();
-    }
-
-    public static void main(String[] args) {
-        Application dialog = new Application();
-        dialog.pack();
-        dialog.setVisible(true);
-        System.exit(0);
-    }
-}
