+++ src/test/java/tudelft/ti2806/pl3/MainTest.java
@@ -1,0 +1,7 @@
+package tudelft.ti2806.pl3;
+
+import junit.framework.TestCase;
+
+public class MainTest extends TestCase {
+
+}
